Cleaned up 2x2 plot function, added data
Updated the 2x2 plot function to use the ellipse package (in concert
with the concurrent ratings plot function). Obviated the need for the
function bivar_norm(), so got rid of that. Cleaned up the rest of the
2x2 plot function to simplify the code and make the figures prettier.
Also added an example confusion matrix to the included data set (from
Silbert (2012)).