siTest, mriTest functions cleaned up
Changed some labels for the siTest and mriTest functions, also added
round(X,3) to the output tables, and corrected a couple minor errors in
the mriTest function. Specifically, made sure the MRI tests are being
carried out on the correct response data, and made sure that the test
statistics are consistent for the A and B tests (with the marginal
response proportion for the second level subtracted from the marginal
response proportion for the first level).